+++ fcm_config/lib/src/web/details.dart
@@ -1,0 +1,15 @@
+class WebNotificationDetails {
+  final String? dir;
+  final String? tag;
+  final String title;
+  final String body;
+  final String? icon;
+
+  WebNotificationDetails({
+    this.icon,
+    required this.body,
+    required this.title,
+    this.dir,
+    this.tag,
+  });
+}