+++ fcm_config/lib/src/web/web_notification_manager.dart
@@ -1,53 +1,53 @@
 library web_forground_notification;
 
 import 'dart:async';
+import 'dart:html' as html;
 
 import 'package:firebase_messaging/firebase_messaging.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
 
-part 'forground_notification.dart';
-
-class WebNotificationDetails {
-  final Duration duration;
-  final String? subTitle;
-  final String title;
-  final String body;
-  final Widget? icon;
-
-  WebNotificationDetails({
-    this.duration = const Duration(seconds: 3),
-    this.icon,
-    required this.body,
-    required this.title,
-    this.subTitle,
-  });
-}
+import 'details.dart';
 
 class WebNotificationManager {
-  static Map<int, OverlayEntry> overlays = {};
   static final StreamController<RemoteMessage> onLocaleClick =
       StreamController<RemoteMessage>.broadcast();
 
-  static void show(BuildContext context, WebNotificationDetails details,
-      [Map<String, dynamic>? data]) {
-    var overlayState = Overlay.of(context);
-    if (overlayState != null) {
-      var id = DateTime.now().microsecondsSinceEpoch;
-      overlays[id] = OverlayEntry(
-        builder: (ctx) => _ForgroundNotificationView(
-          details: details,
-          id: id,
-          onClick: () {
-            if (data != null) onLocaleClick.add(RemoteMessage.fromMap(data));
-          },
-          onDismiss: (_id) {
-            overlays[_id]?.remove();
-            overlays.remove(_id);
-          },
-        ),
-      );
-      overlayState.insert(overlays[id]!);
+  static Future show(WebNotificationDetails details,
+      [Map<String, dynamic>? data]) async {
+    var permission = html.Notification.permission;
+    if (permission == 'default') {
+      permission = await html.Notification.requestPermission();
     }
+    if (permission != 'granted') return;
+    var notifification = html.Notification(
+      details.title,
+      body: details.body,
+      icon: details.icon ?? '/icons/icon-192.png',
+      dir: details.dir,
+      tag: details.tag,
+    );
+    var notify = RemoteMessage(
+        collapseKey: details.tag,
+        data: data ?? {},
+        from: 'locale',
+        sentTime: DateTime.now(),
+        contentAvailable: true,
+        notification: RemoteNotification(
+          title: details.title,
+          body: details.body,
+          android: AndroidNotification(
+            smallIcon: details.icon,
+          ),
+          apple: AppleNotification(
+            imageUrl: details.icon,
+          ),
+        ));
+    var click = notifification.onClick.listen((event) {
+      onLocaleClick.add(notify);
+    });
+    StreamSubscription<html.Event>? close;
+    close = notifification.onClose.listen((event) {
+      click.cancel();
+      close?.cancel();
+    });
   }
 }